Split exports under nets into individual files by purpose: API unchanged.
